requirements.txt
beautifulsoup4
seleniumbase
python-telegram-bot
customtkinter
dnspython
